Removed a few pedantic warnings.

+++ extras/JuceDemo/Source/demos/AudioDemoRecordPage.cpp
@@ -120,7 +120,7 @@ public:
         // We need to clear the output buffers, in case they're full of junk..
         for (int i = 0; i < numOutputChannels; ++i)
             if (outputChannelData[i] != 0)
-                zeromem (outputChannelData[i], sizeof (float) * numSamples);
+                zeromem (outputChannelData[i], sizeof (float) * (size_t) numSamples);
     }
 
 private:
