+++ modules/juce_core/files/juce_File.h
@@ -580,7 +580,7 @@ public:
                     end of the file), or nullptr if the file can't be opened for some reason
         @see createInputStream, appendData, appendText
     */
-    FileOutputStream* createOutputStream (int bufferSize = 0x8000) const;
+    FileOutputStream* createOutputStream (size_t bufferSize = 0x8000) const;
 
     //==============================================================================
     /** Loads a file's contents into memory as a block of binary data.
