+++ modules/juce_core/zip/juce_GZIPCompressorOutputStream.cpp
@@ -103,7 +103,7 @@ private:
                 {
                     data += dataSize - stream.avail_in;
                     dataSize = stream.avail_in;
-                    const ssize_t bytesDone = sizeof (buffer) - (ssize_t) stream.avail_out;
+                    const ssize_t bytesDone = (ssize_t) sizeof (buffer) - (ssize_t) stream.avail_out;
                     return bytesDone <= 0 || out.write (buffer, (size_t) bytesDone);
                 }
